+++ src/main/java/module-info.java
@@ -3,22 +3,16 @@ open module maven.fetcher {
 
     exports maven.fetcher;
 
-    requires aether.api;
-    requires aether.connector.basic;
-    requires aether.impl;
-    requires aether.spi;
-    requires aether.transport.file;
-    requires aether.transport.http;
-    requires aether.util;
+    requires org.apache.maven.resolver;
+    requires org.apache.maven.resolver.connector.basic;
+    requires org.apache.maven.resolver.impl;
+    requires org.apache.maven.resolver.spi;
+    requires org.apache.maven.resolver.transport.file;
+    requires org.apache.maven.resolver.transport.http;
+    requires org.apache.maven.resolver.util;
     requires org.apache.commons.lang3;
+    requires maven.aether.provider;
     requires transitive org.slf4j;
-    requires transitive java.instrument;
-    requires plexus.utils;
-    requires plexus.component.annotations;
-    requires plexus.interpolation;
-    requires javax.inject;
-    requires com.google.guice;
-    requires com.google.common;
     requires slf4jansi;