+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
+## [1.1.0]
+**Release date:** 2021/05/11
+
+### Modified
+- Use of Apache Maven Artifact Resolver instead of Aether components 
+  as underlying implementation, as referred in https://github.com/luiinge/maven-fetcher/issues/10
+
 ## [1.0.0]
 **Release date:** 2020/07/04
 
@@ -13,4 +20,4 @@ and this project adheres to [Semantic Versioning][2].
 
 
 [1]: <https://keepachangelog.com/en/1.0.0/>
-[2]: <https://semver.org/spec/v2.0.0.htm>
+[2]: <https://semver.org>